Find nearest address in NLSNearestFeatureSearchChannel using NLSFIGeocodingSearchChannel
